requirements.txt
asgiref==3.5.2
autopep8==2.0.0
Django==2.2.28
django-filter==22.1
djangorestframework==3.14.0
Markdown==3.4.1
psycopg2==2.9.3
pycodestyle==2.9.1
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.6
